feat(Android): PoC of tabs for Android (#111)
## Description

This is a PoC of implementation of "bottom tabs" concept for Android
platform.
I went here for usage of
[`BottomNavigationView`](https://m3.material.io/components/navigation-bar/overview)
and not [`Tabs`](https://m3.material.io/components/tabs/overview). I
consider `BottomNavigation` view to better reflect the idea we want to
implement here. At the same time, I want to emphasise that *Material
Tabs* will be implemented in the future, as separate component.

To simplify the setup I used already existing native components, but in
the future this will be refactored according to the guidelines
https://github.com/software-mansion/react-native-screens-labs/discussions/107.

This PR introduces only a base skeleton of the implementation, with
functionalities left for followups (similar process as for iOS tabs).

Please see the commits & their descriptions for details on particular
changes.

### Fragments

I went for tab management using fragments, because it allows for easy
access to screen lifecycle. W/o it it might get tricky to strictly
define on{Will,Did}{Appear,Disappear} moments, however this is something
to consider knowing fragments pitfalls.

## Test code and steps to reproduce

## Checklist

- [ ] Ensured that CI passes