feat(Android): emit `onNativeFocusChange` & react to JS focus change (#116)
This PR adds required events & logic for basic tab switch to work. 

Currently, I assume that the component runs in JS-controlled mode & it
is
the only scenario supported.

On signal of native tab change, we send an event to JS & wait for next
render.
Upon receive of focus change signal (prop update) we trigger container
update
(very similar to the architecture on iOS).


---


- **Add `TabsHostNativeFocusChangeEvent`**
  

- **Add `TabsHostEventEmitter`**
  

- **Setup container update on JS focus change mechanism**
  I really don't like that the container logic resides in `TabsHost`.
We should refactor it out in some later step & make sure it's separated
from components logic. For the sake of making it work I want to proceed
  now, however it will be refactored in follow-ups.