refactor(Android): structurize updates to tab container & run them on as-needed basis (#123)
## Description

This PR adds `ContainerUpdateCoordinator`.

This class is meant to coordinate updates to the tab container.
It'll avoid sending tons of messages to main looper as it is in current
approach & will enable us to update only what's needed at given moment,
w/o updating always all the aspects of the container.

Since this is a PoC, the implementation lives currently as inner class
in
`TabsHost`, but it should be later moved to standalone class.

This refactor also fixes #122

During research it also came up that we can register for
`mountingTransaction{Will,Did}Mount` events on `FabricUIManager` &
therefore mimic behaviour from iOS. However I defer that to follow-up
PR.


## Test code and steps to reproduce

`TestBottomTabs`

## Checklist

- [ ] Ensured that CI passes