feat(iOS, Tabs): add SFSymbol icons, iconColor and titlePositionAdjustment (#104)
Adds `iconSFSymbolName`, `selectedIconSFSymbolName`,
`tabBarItemIconColor` and `tabBarItemTitlePositionAdjustment`.

1. I created seperate props for regular icon and selected icon but did
not do the same for `tabBarItemIconColor`. This is because icon itself
is set on UITabBarItem directly and icon color is set as a part of the
appearance. In the future, we will separate appearance into states for
all appearance props.
2. I tried to use `imageInsets` for the icon but on iOS 18 this does not
seem to be working so I abandoned it.
3. You can change icon size with `UIImageSymbolConfiguration` but you
can only apply really small changes - if you make the icons too big,
they just overflow and the bar does not adapt. If we still want to allow
for size customization (e.g. to make the icons a little bit
larger/smaller), we can do that in this PR or in the future.
4. I used regular string as type for SFSymbol icon name but we can
consider using `sf-symbols-typescript`.

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/93
as `imageInsets` does not seem to be working.
Closes
https://github.com/software-mansion/react-native-screens-labs/issues/95.