feat(iOS): new three-state lifecycle for stack (#148)
## Description

This PR follows discussion in #139, implementing three-state lifecycle
state model
for stack screens.

Supersedes #141

Now, each screen can be in one of three states: `INITIAL`, `DETACHED`,
`ATTACHED`, with meaning as follows:

* `INITIAL` - the screen has not yet been rendered, the native side is
not aware of its existence
* `DETACHED` - the component has been rendered at least once & there
exists native instance; component in this state can not be attached to
**native controller**,
* `ATTACHED` - the component has been rendered at least once & there
exists native instance; component in this state will be/is attached to
**native controller**.

The state of the component is controlled via `maxLifecycleState` prop.
E.g. if you render a `StackScreen` with `maxLifecycleState ===
DETACHED`,
you will end up with rendered component, retained on native side, but
not attached to `UINavigationController`.

To dismiss a component it is enough to reduce the state to `DETACHED`.
Upon dismissal it can be decided in JS code whether to stop rendering a
screen or keep it ready (potentially preloaded).

For broader discussion of semantics please see #139.

This PR additionally does some small renamings & modifies the stack
example.
Now, repeated navigation to screen with the same name has no effect &
screen keys are required to be passed up front.
I'm not certain this will be our final API, however I wanted to align it
with current bottom tabs approach.

Note, that in current implementation only dismissal of the top-screen is
possible. We also need to add `onDismissed` event,
so that it is clear whether a native screen has been dismissed or not. 

Another important note is that preload is not yet supported*, since the
screens are pushed onto the stack in rendering order.

Another note: native back is not supported; this is not a regression - I
believe it didn't work correctly earlier, however it became apparent
only after the aforementioned change to `push` behaviour. It will be
easy to solve, once we introduce `onDismissed` event.

## Test code and steps to reproduce

`TestStackScreen`

## Checklist

- [ ] Ensured that CI passes