feat(iOS, Tabs): add support for popping nested stack to root on repeated tab selection (#154)
## Description

Add support for popping nested stack to root on repeated tab selection.

In UIKit, this behavior is already implemented and enabled by default
but UINavigationController has to be a direct child of
UITabBarController. In our implementation, there is a
RNSTabsScreenViewController between them - that's why we need to
implement this interaction ourselves. In the future we might consider
implementing special tab screen component that will serve as a
navigation controller as well.

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/170.

## Changes

- add `RNSBottomTabsSpecialEffectsSupport` protocol
- use new interaction with `RNSNavigationController`
- add Tab4 to test it 

## Test code and steps to reproduce

Enter Tab4, go to Screen2/Screen3 and click on Tab4 again.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes