feat(iOS, SplitView) Setup basic logic for custom ShadowNodes (#172)
## Description

This PR is adding a basic setup for `RNSSplitViewScreenShadowNode` which
is responsible for aligning (x, y) offsets in Shadow tree relatively to
native layout metrics.

This PR is also adding another controller
`RNSSplitViewNavigationController`. According to the doc:
https://developer.apple.com/documentation/uikit/uisplitviewcontroller#Child-view-controllers
- we can wrap our components in custom NavigationController. In our case
`RNSSplitViewNavigationController` is responsible for attaching
`NSKeyValueObservation` for `frame`. It's necessary to track updates
this way, as `layoutSubviews` won't be called when we're showing or
hiding columns, because their width remains constant, no layout update
is needed and only origin for our component (and its children
relatively) is updated by some (x, y) vector.

## Changes

- Updated SplitViewBaseApp
- Added necessary logic for new ShadowNode in common directory 
- Updated SplitView native implementation to read and pass layout to
ShadowNode

## Test code and steps to reproduce

At the moment, TestSplitView is covering all cases, for testing
purposes. Until we have some props supported, I'd recommend overriding
SplitView props directly on the native side in
`RNSSplitViewHostController`

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

Closes: #180

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>