feat(Android,Tabs): add item tittle color props (#192)
## Description

Resolves: #204

This PRs adds possibility to set item title color for all tabs on
Android. It's set on TabHost cause we set it later also on
`bottomNavigationView`. It could later added on the TabSceen as well to
match iOS, although when testing it wasn't possible to set different
colors on each item.

## Test code and steps to reproduce

See `TestBottomTabs`

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Ensured that CI passes