feat(iOS, SplitView): Adapt SplitView layout integration to iOS 26 requirements (#221)
## Description

This PR is introducing changes in the ShadowTree layout for out
SplitView, adapting to the following iOS26 changes:
- SplitView columns are resizeable, we need to additionally track layout
updates.
- A new `inspector` column was added - the purpose for this column is
presenting some additional details for the content shown in a
`secondary` column. This column has a different presentation type, based
on the screen orientation and SizeClass. If there's enough space - this
column on a trailing edge for the secondary column. If not, then it's
shown as a Modal component,
- `inspector` in modal-style is positioned outside the RN hierarchy, so
we need to add some more adjustments, like attaching touch handler

Additionally, as we discussed, we want to have a `RootNodeKind` trait
for our SplitViewScreen ShadowNode in both scenarios for rendering
column or modal.

## Changes

- added `RootNodeKind` trait to `RNSSplitViewScreenShadowNode.h`
- attaching touch handler if we recognized that the column is presented
in modal-style
- overriding `viewDidLayoutSubviews` for iOS26 to handle dynamic
resizing for columns

## Test code and steps to reproduce

Updated `SplitViewBaseApp`.

Notes:
- I'm leaving some TODOs that I'm planning to remove later, because as
for now, toggling the inspector can be done only programmatically
calling `show(_ column: UISplitViewController.Column)`.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

Closes: #181