feat(legacy-arch): basic support for legacy architecture for bottom tabs (#273)
## Description

This PR adds support for legacy architecture for bottom tabs component
on both main platforms.

The working principle of containers is a bit different on Paper, namely
containers do not wait for flush signal (on Fabric this is
`reactMountingTransactionDidMount`), but rather eagerly **schedule** an
update block on each update request. The block scheduling is (partially)
indempotent - only a single block will be scheduled at a given time.

The rest of the logic is mostly shared, with few exceptions, such as
prop updates & event emitting code, which required separate
implementations.

## Test code and steps to reproduce

`TestBottomTabs` should work just fine in `Example` (remember to install
pods with `RNS_GAMMA_ENABLED=0`, otherwise it might fail to build).

## Checklist

- [ ] Ensured that CI passes

---------

Co-authored-by: kligarski <63918941+kligarski@users.noreply.github.com>