feat(iOS, Tabs): add support for custom images and templates for tab bar item icons (#3007)
Add support for custom icons in tab bar items.

There were many ideas how to handle sizing of the icons. With
@maciekstosio, we decided that the best approach for now is to mimic how
`Image` from `react-native` works. If we want to use sharp-looking
icons, we need to provide the icon for different screen densities, more
details
[here](https://reactnative.dev/docs/images#static-image-resources).

| 2x (iPhone SE) | 3x (iPhone 16 Pro) |
| -- | -- |
| <video
src="https://github.com/user-attachments/assets/57714811-82cc-47ef-9d09-a855198788ea"
/> | <video
src="https://github.com/user-attachments/assets/011dc2ee-0513-4dce-8d3a-cbdcb4e462c8"
/> |

For now, in order to change the icon size, you need to scale image files
properly. Ideal sizes are available in
[docs](https://developer.apple.com/design/human-interface-guidelines/tab-bars#Target-dimensions).
Images are not scaled by UIKit in stacked layout - this is how it works
on native platform (SwiftUI as well). What is interesting is that in
compact mode (e.g. horizontal orientation) the images are scaled but I
wasn't able to confirm this behavior with the UIKit docs.

In the future, we might consider providing a property to scale the image
(like `style: {width, height}` for `Image` but we need to consider how
to implement this) and expose `imageInsets` - they work with icon in
stacked layout and might be useful when using custom font.

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/92.

Moved from
https://github.com/software-mansion/react-native-screens-labs/pull/142.