Add stubs for missing symbols when building with RNS_GAMMA_ENABLED=0
These symbols are a stubs for components defined in Gamma project implementation.
We need these, because we can not really trick codegen to generate code only for components
that satisfy some condition (e.g. when env var is defined). If these symbols are missing,
the application will fail in runtime when RN attempts to create components class registry.