fix(iOS): provide missing symbols when building w/o Gamma project enabled (#3032)
## Description

Closes #3031

This one is on me... I've fogotten that I've defined
`RNS_GAMMA_ENABLED=1` on my machine &
apparently I've tricked myself when testing :/ 

These symbols are a stubs for components defined in Gamma project
implementation.
We need these, because we can not really trick codegen to generate code
only for components
that satisfy some condition (e.g. when env var is defined). If these
symbols are missing,
the application will fail in runtime when RN attempts to create
components class registry.

If there is a method to include only a subset of components into code
generation process depending on a condition, we should use that, however
I'm not aware of such possibility atm.

## Changes

Added stubs for missing symbols when building with `RNS_GAMMA_ENABLED=0`
or undefined.

## Test code and steps to reproduce

Runtime should not crash...

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes