feat(Android, Tabs): support badge in bottom navigation (#3017)
## Description

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/256

This PRs adds basic support for badge for android. It allows to set
color and label. To reuse the iOS logic (which allows string as badge
value) I need to parse it to int and I log error otherwise. ~This isn't
ideal and I created low-priority task to change it at some point
https://github.com/software-mansion/react-native-screens-labs/issues/249.~

I merged rest of the PRs related to badge to this PR:
- text color 
- label visibility
- handling text 

It currently works in a way that if it can parse it as int, it will be a
number and displayed as string otherwise. Developer can use just label
visibility to show "badge dot". At the same time I got rid of the number
warning. It could be potentially changed to have separate methods for
string and number, but for now I conformed to iOS API.

## Screenshots / GIFs

<img width="436" height="928" alt="image"
src="https://github.com/user-attachments/assets/cb9f6b0b-8953-4061-a47f-2478842ef321"
/>

## Test code and steps to reproduce

See `TestBottomTabs/index.tsx`

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Ensured that CI passes