fix(iOS, Gestures): Ignore inactive preloaded screens in gestureRecognizer (#3045)
## Description

Fixes https://github.com/react-navigation/react-navigation/issues/12683

Currently, `gestureRecognizer:shouldReceivePressOrTouchEvent` decides
whether it wants to handle touch events by checking properties of the
last screen in subviews list. It doesn't check if the last screens is
actually the one that is presented. The bug in
https://github.com/react-navigation/react-navigation/issues/12683 came
from the fact, that preloaded screen was added to the end of the list
and was taken as the `topScreen`.

## Changes

To find the `topScreen`, the subview list is iterated backwards to find
the first screen that has state=OnTop. When new screens are preloaded to
the end of the list, they have different state, and are ignored.

## Test code and steps to reproduce

Use the provided Test.