feat(iOS, SplitView): Add API for organizing columns (#3003)
## Description

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/229

In this PR, I'm adding new APIs for SplitView column management. I
updated `SplitViewScreen` component export to separate `Column` which is
rendered as typical SplitViewColumn and `Inspector` which in some cases
might be rendered as modal. Additionally, I'm adding a prop to Host for
toggling inspector, at the moment it's really laggy, but I'm leaving it
as a followup.

In this PR we also noticed that the logic for updating props needs to be
updated. For number of columns management we need to delay
`SplitViewHostController` initialization until we'll receive
`reactSubviews` from the initial render. This is causing a problem, when
1st `updateProps` call is coming, because if inspector should be
displayed, we need to notify the component about that fact. Therefore,
this logic was rewritten to collect all prop updates in batch, create
controller on 1st `updateProps` call and notify controller to flush
updates on `didMount`.

## Changes

- Updated `SplitViewBaseApp`
- Changed the logic for props updates
- Added `showInspector` prop to host component
- Added `columnType` prop to screen component
- Added assertions for unexpected column manipulations
- Updated JS definitions

## Test code and steps to reproduce

Updated `SplitViewBaseApp`


https://github.com/user-attachments/assets/22f65d96-6623-4e70-9938-182a36ed7a9d


https://github.com/user-attachments/assets/c0608475-b26e-4cfd-b62f-99911d813beb

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes