feat(iOS, SplitView) Add support for displayModeButtonVisibility prop (#3010)
`displayModeButtonVisibility` - A setting that determines whether the
display mode button is visible in the interface.

During testing, I found that layout is broken for display mode with 3
columns open from start, therefore I'm fixing it updating the
ShadowState with the frame that I received in `viewDidLayoutSubviews`.

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/232

## Changes

- Added `displayModeButtonVisibility` prop
- Fixed layout for multiple column display

## Screenshots / GIFs


https://github.com/user-attachments/assets/01d311ed-9585-4e82-b09e-5d0dc365e5d6

## Test code and steps to reproduce

Updated app to cover this prop

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes