fix(iOS, SplitView): Fix SplitViewHostController initialization (#3054)
## Description

Initializing Host controller on `didMoveToWindow` is causing in losing
some updates when the appearance coordinator was called earlier from
`finalizeUpdates`. Props on the native side were already set, the native
state was invalidated, but the appearance coordinator was skipped,
because controller wasn't initialized at this point.

## Changes

- Updated a logic to consume appearance updates only after the
controller was initialized.

## Test code and steps to reproduce

Verified that the current app sets all props properly.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes