fix(iOS, SplitView): Update showSecondaryToggleButton dynamically (#3047)
## Description

This PR adds an option to update `showSecondaryToggleButton` prop
dynamically. As for now, it requires some layout update to trigger
refresh of the navbar to which this button is attached. After some
investigation, we found that we can forcefully update that navbar by
hiding it and opening again. With an assumption that it's done in a
single transaction, no flicker/lag is observed.

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/285
.

## Changes

- Added native flag for signaling that secondary screen navbar update is
needed.
- Added logic for triggering updates for secondary navbar.

## Screenshots / GIFs


https://github.com/user-attachments/assets/998a1189-368b-4e42-8dae-7930b02e7874

## Test code and steps to reproduce

Updated SplitViewBaseApp

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes