feat(iOS, SplitView): Add support for fraction for columns widths (#3043)
## Description

In this PR I'm covering UISplitViewController APIs
`preferred[ColumnType]ColumnWidthFraction`. As we discussed internally,
we should try to cover it as the same prop for both fraction and points.
This is slightly different from the Apple approach which has both
versions separated, maintaining two props:
`preferred[ColumnType]ColumnWidth` and
`preferred[ColumnType]ColumnWidthFraction`.

In our case, if the value passed is between 0 and 1 - we're interpreting
it as a fraction and for any value greater than 1, we're using points.

Closes:
https://github.com/software-mansion/react-native-screens-labs/issues/296

## Changes

- Updated `preferredColumnWidth` definitions to cover floats.
- Updated appearance coordinator to cover fractions.
- Updated prop definition in the host component.

## Screenshots / GIFs


https://github.com/user-attachments/assets/07ef660c-f74e-4bc8-b855-8d197e4db840

## Test code and steps to reproduce

Updated example for SplitView for testing this API

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes