fix(iOS, SplitView): Add flag for invalidating preferredDisplayMode (#3051)
## Description

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/295

displayMode chosen by OS might differ from the `preferredDisplayMode`
value that we keep on the JS side. This may cause a flicker, when user
is updating dynamically any other prop on the Host component, because,
as for now, `preferredDisplayMode` would be set once again and will be
synchronized with the JS value, forgetting about all native interactions
that resulted in changing state.

In this PR, I'm adding some extra logic to prevent that case:
- update `preferredDisplayMode` only when JS prop was changed

## Changes

- Update appearance coordinator and SplitViewHost logic for updating
props

## Screenshots / GIFs

### Before


https://github.com/user-attachments/assets/ca5436f6-7742-41c8-9826-26c0984ff026

### After


https://github.com/user-attachments/assets/fd31d5f0-204e-4594-bfd8-c39dddf7bce1

## Test code and steps to reproduce

Updated demo app

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes