feat(Example): Fix sorting for multiple tests with same issue number (#3067)
## Description

Currently, we write tests as issue reproductions which we name
Test<issue number> - one for each issue, and more general tests/examples
which we name as Test<descriptive name>. In order to have the former
sorted correctly, we need to stick to the scheme and not have anything
non-digit after issue number.

We can change the sorting logic to make it possible to define Test123_1
and Test123_2.

## Changes

Implementation of sort function that understands names like Test123_1