fix(Android): Differentiate between new and preloaded screens in ScreenStack (#3062)
## Description

Closes #3004.

When `onUpdate` is issued for ScreenStack, we have no way of knowing
whether the screen that is on top with state=2 was already on top or was
preloaded with state=0, so we need to store this information somehow,
i.e in a list.

## Changes

Added `preloadedWrappers` list to `ScreenStack` class and update it at
the end of each `onUpdate`.


## Test code and steps to reproduce

See Test3004