fix(iOS, SplitView): Break the retain cycle between component and controller for SplitViewScreen (#3083)
## Description

Closes:
https://github.com/software-mansion/react-native-screens-labs/issues/156

Problem:
- Controller has a strong reference to an associated component via
`.view` property.
- In our case, Component keeps a strong reference to the controller.
- This creates a retain cycle, which we're not breaking anywhere at the
moment.

Solution:
SplitViewHost keeps references to all subviews of type SplitViewScreen
all the time (columns might be hidden, but they won't disappear). It
means that we can consider Host as an owner of the SplitViewScreens'
lifecycles. If the host is about to be unmounted from the view
hierarchy, we can perform some additional cleanup in child components to
detach the reference to theirs controllers.

## Changes

- Added a method for invalidating a screen component that removes a
reference to the controller.

## Test code and steps to reproduce

Create an example that includes a SplitView, which may be removed from
the hierarchy, add log and observe.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacper.kafara@swmansion.com>