feat(iOS): add an option to prevent dismissal of external modals (#2994)
## Description

In some cases, dismissing a modal not owned by RNScreens may not be the
expected behavior.

## Changes

Added a new protocol, `RNSDismissibleModalProtocol`, which can be used
by external libraries to control whether their modal should be dismissed
when pushing a screen’s modal on top.

When dismissing the top modal, the code now checks for conformance to
this protocol. If the modal implements the protocol and `isDismissible`
is `true`, the dismiss will not happen.

In all other cases, the behavior of `setModalViewControllers` remains
the same as before this change.

## Test code and steps to reproduce

<!--
Please include code that can be used to test this change and short
description how this example should work.
This snippet should be as minimal as possible and ready to be pasted
into editor (don't exclude exports or remove "not important" parts of
reproduction example)
-->

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes