chore(Lint): Run swift lint on macos runner, split jobs between runners, remove formatting step (#3088)
## Description

Currently, GH action for installing swift on ubuntu crashes randomly. In
this PR, swift-format is moved to run on macos vm, and other linters are
moved to separate jobs on ubuntu runner.

Additionally, `yarn format` is removed because it doesn't return errors,
just formats files in runner's filesystem, and it also runs swift-format
which won't be available on ubuntu runners.

## Changes

- Lint github workflow
- yarn lint-swift uses xcrun instead of swift-format directly
- `yarn format` is removed