Use `target_compile_reactnative_options` to prevent weird crash
The function has been introduced with RN 0.80 and later modified in
https://github.com/facebook/react-native/commit/c059ae1b77.

The #c059ae1b77 added RN_SERIALIZABLE_STATE compile time definition,
which helps to resolve an ridiculous crash / hang we had. To give some
context - the crash occured in codegened C++ code in
`RNSScreenProps::~RNSScreenProps` (yes, this is a compiler generated
default dtor) or in other dtors, depending on exact timing. It is still
unknown exactly what was the error mechanism, but most likely due to
some timing / concurrency related issues dtor attempted to free some
memory that was no longer mapped to a process or something. Anyway,
setting this flag resolves the problem.

Additionally, this is also a way-to-go in terms of setting compile
options required by RN since 0.80.