Add workaround for id of TabsHost colliding with ReactSurfaceView
`TabsHost.id` collied with `ReactSurfaceView.id`, because RN 0.81
started to number the root views from 1, instead of 11 as it was before.
This caused issues, becuase `FragmentTransaction.add(containerViewId:
Int, ...)` requires us to pass the `containerViewId`, which it then uses
to find a view in hierarchy with that id & it mounts added fragment's
view there. If collision occurs, Android Framework mounts the fragment
directly under `ReactSurfaceView` instead of our `TabsHost`.

Proposed workaround does the following:

1. it passes int value of react native minor version down from gradle
   script to build config, so that it is accessible in platform code.
   AFAIK there is no dedicated API as of 0.81 for this, therefore we
   need a workaround.
2. It adds new `ViewIdGenerator` that implements new `ViewIdProviding`.
   It has capability of overriding the view generator, so that the lib
   can be integrated in brownfield scenarios. We should use this object
   to generate view id for any of our views that need it.
3. The `ViewIdGenerator` relies on internals of RN, that are referenced
   in the code comments, to provide non-colliding view ids (tags).