fix(iOS): Glitching back navigations on iOS 26 (#3093)
## Description, Changes

This PR handles the change in behavior of screen navigation using back
button, which, on iOS 26, doesn't wait for one transition to finish and
can pop two screens at once - something we are not able to handle
correctly. In order to do that, the back button is set to be disabled
while transition is happening.

The `rnscreens_findBackButtonWrapperView` method from UINavigationBar
extension, which was needed for finding the back button in navigation
bar, is fixed for iOS 26. The method has been used to calculate insets
for shadow tree and must have been failing for iOS 26 because the button
has changed its location and class. This likely fixes another
to-be-found bug.

iOS 26 also introduces new `interactiveContentPopGestureRecognizer`
which can fire mid transition. Handling the new recognizer is beyond the
scope of this PR so it is set to be disabled permanently in a followup
PR.

### Before


https://github.com/user-attachments/assets/489b0fd6-2e9a-41cb-8e05-2053de179106

### After


https://github.com/user-attachments/assets/0ba706de-58b4-4373-91fc-7c3db9e4d651


## Test code and steps to reproduce

See the provided test case.

---------

Co-authored-by: Kacper Kafara <kacper.kafara@swmansion.com>