fix(example): Make react-freeze work in BottomTabs (#3116)
## Description

To merge after #3115 

Fixes
https://github.com/software-mansion/react-native-screens-labs/issues/354.

Currently, BottomTabsContainer supports passing a function that will
return ReactNode and injects it inside its markup, inside <Freeze>
component like so: `<BottomTabsContainer> { tabs.map( () => <Freeze> {
componentFunction() } </Freeze> } </BottomTabsContainer>`. This way, it
too late for Freeze to trigger, since the function has completed and
rendered everything it was supposed to. The PR no longer accepts
functions and requires passing a React.Component, which will be rendered
only after it was passed as `children` prop into `<Freeze>`

The function to be removed had one parameter that was passed a
`selectNextTab` function, but we came to a conclusion that its a
leftover and no longer needed.

## Changes

Changed TabConfiguration.contentViewRenderFn to
TabConfiguration.component. Changed returned value from function to
React.ComponentType. Removed dead code.

## Test code and steps to reproduce

Use Test3115. First screen has freezing disabled, second has it enabled.
Observe counter values incrementing for first screen when second is
focused, and not incrementing for second when first is focused.