fix(iOS): Disable second transition coming from gestures when one is in progress (#3142)
## Description

This is a followup to #3093 and #3141. We now handle repeatedly clicking
back button on iOS 26, and disable
interactiveContentPopGestureRecognizer which triggers without
fullScreenGesture prop enabled, but it still is possible to click back
button and swipe when fullScreenGesture is enabled. One possible
solution is to check if transitionCoordinator is set and abort handling
the gesture if it is.

## Changes

Added check for transitionCoordinator == nil before allowing for gesture
to be received.

## Test code and steps to reproduce

Use Test3093.