chore: Hide native logs (#3144)
## Description

When using bottom tabs, many logs are printed from the native
components. These logs are useful for development and debugging in
screens library, but are mostly noise when using in external app

## Changes

Adds `RNSLog` function on iOS and `RNSLog` object on Android. Both can
be conditionally enabled for debugging.

## Test code and steps to reproduce

Install iOS pods with `RNS_DEBUG_LOGGING=1 pod install` to see the logs
on iOS.

For Android `rnsDebugLogsEnabled` needs to be set to `true` in
`gradle.properties`

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes